[HN-205] refactor: Use inheritance for LLM classes (#27)
* Using inheritance in ClaudeLLM

* Using inheritance in OpenAILLM

* Using inheritance in MistralLLM

* Using inheritance in OllamaLLM

* Using llms.py as super class
