+++ hive_agent/llms/__init__.py
@@ -2,3 +2,4 @@ from .openai import OpenAILLM
 from .claude import ClaudeLLM
 from .ollama import OllamaLLM
 from .mistral import MistralLLM
+from .llms import LLMs
