+++ hive_agent/llms/ollama.py
@@ -1,28 +1,8 @@
 from llama_index.core.settings import Settings
 from llama_index.core.agent import FunctionCallingAgentWorker
 
-class OllamaLLM:
-    def __init__(self, tools, instruction):
-        self.tools = tools
-        self.instruction = instruction
-        self.system_prompt = f"""You are a domain-specific assistant that is helpful, respectful and honest. Always 
-                answer as helpfuly as possible, while being safe. Your answers should not include any harmful, 
-                unethical, racist, sexist, toxic, dangerous, or illegal content. Please ensure that your responses are 
-                socially unbiased and positive in nature. If a question does not make any sense, or is not factually 
-                coherent, explain why instead of answering something not correct. If you don't know the answer to a 
-                question, please don't share false information.
+from hive_agent.llms.llms import LLMs
 
-                You may be provided with tools to help you answer questions. Always ensure you use the result from 
-                the most appropriate/relevant function tool to answer the original question in the user's prompt and comment
-                on any provided data from the user or from the function result. Whenever you use a tool, explain your answer
-                based on the tool result.
-
-                Here is your domain-specific instruction:
-                {self.instruction}
-                """
-        self.agent = FunctionCallingAgentWorker.from_tools(
-                tools,
-                system_prompt=self.system_prompt,
-                llm=Settings.llm,
-                allow_parallel_tool_calls=False,
-                ).as_agent()
+class OllamaLLM(LLMs):
+        pass   
+
