+++ hive_agent/llms/openai.py
@@ -1,23 +1,7 @@
 from llama_index.agent.openai import OpenAIAgent
+from hive_agent.llms.llms import LLMs
 
-class OpenAILLM:
+class OpenAILLM(LLMs):
     def __init__(self, tools, instruction):
-        self.tools = tools
-        self.instruction = instruction
-        self.system_prompt = f"""You are a domain-specific assistant that is helpful, respectful and honest. Always 
-                answer as helpfuly as possible, while being safe. Your answers should not include any harmful, 
-                unethical, racist, sexist, toxic, dangerous, or illegal content. Please ensure that your responses are 
-                socially unbiased and positive in nature. If a question does not make any sense, or is not factually 
-                coherent, explain why instead of answering something not correct. If you don't know the answer to a 
-                question, please don't share false information.
-
-                You may be provided with tools to help you answer questions. Always ensure you use the result from 
-                the most appropriate/relevant function tool to answer the original question in the user's prompt and comment
-                on any provided data from the user or from the function result. Whenever you use a tool, explain your answer
-                based on the tool result.
-
-                Here is your domain-specific instruction:
-                {self.instruction}
-                """
-        print("openAI LLM initialized")
+        super().__init__(tools, instruction)
         self.agent = OpenAIAgent.from_tools(tools=self.tools, system_prompt=self.system_prompt)
